+++ ChangeLog.md
@@ -10,6 +10,7 @@ Blob:
 
 - Fixed issue of filtering blobs with correct multiple conditions on single tag (range queries). (issue #2514)
 - Added support for sealing append blobs. (issue #810)
+- Added support for delegation sas with version of 2015-07-05.
 
 Table: